+++ CHANGELOG.md
@@ -41,8 +41,11 @@
 ### Improvements
 
 - Optimize the performance of scanning incompatible characters.
+- Accept the JSON5 format when importing custom themes and multiple replace definitions.
 - Improve the behavior of trailing whitespace trimming.
+- Update the JSON syntax to highlight comments in the JSON5 format.
 - Improve stability.
+- [dev] Update the build environment to Xcode 26.3-rc.
 
 
 ### Fixes